requirements.txt
openai~=0.28.1
langchain~=0.0.330
django~=4.2.7
djangorestframework~=3.14.0
django-cors-headers~=4.3.0
pytesseract~=0.3.10
pdf2image==1.16.3
drf-yasg~=1.21.7
psycopg2-binary~=2.9.9
djangorestframework-simplejwt==5.3.1
gunicorn==20.1.0
pinecone-client~=4.0.0
tiktoken~=0.6.0